Move SpyIdBrokerClient to its own file
